refactor(bot): add discord utils barrel
