+++ apps/bot/src/discord/interactionRouter.ts
@@ -11,7 +11,7 @@ import { buildRegistryWithUniqueNames } from '../shared/build-registry.js';
 import { CUSTOM_ID_SEPARATOR } from './constants.js';
 import { AppError, ValidationError } from './errors.js';
 import type { ButtonHandler } from './types.js';
-import { buildOpEmbed } from './utils/build-op-embed.js';
+import { buildOpEmbed } from './utils/index.js';
 
 const allButtonHandlers: Array<ButtonHandler> = [
   ...infoButtonHandlers,
