+++ apps/bot/src/domains/_dev/commands/debug/player.ts
@@ -1,5 +1,5 @@
 import type { Command } from '../../../../discord/types.js';
-import { getTargetUser } from '../../../../discord/utils/get-target-user.js';
+import { getTargetUser } from '../../../../discord/utils/index.js';
 import { findOrCreatePlayer } from '../../../player/service.js';
 
 import { replyDebugData } from './utils.js';
