+++ apps/bot/src/domains/_dev/commands/debug/ship.ts
@@ -1,6 +1,6 @@
 import { NotFoundError } from '../../../../discord/errors.js';
 import type { Command } from '../../../../discord/types.js';
-import { getTargetUser } from '../../../../discord/utils/get-target-user.js';
+import { getTargetUser } from '../../../../discord/utils/index.js';
 import { findOrCreatePlayer } from '../../../player/service.js';
 import * as shipRepository from '../../../ship/repository.js';
