+++ apps/bot/src/domains/_dev/commands/emoji.ts
@@ -2,8 +2,7 @@ import { ActionRowBuilder, ButtonBuilder, ButtonStyle } from 'discord.js';
 import sampleSize from 'lodash/sampleSize.js';
 
 import type { Command } from '../../../discord/types.js';
-import { buildCustomId } from '../../../discord/utils/build-custom-id.js';
-import { buildOpEmbed } from '../../../discord/utils/build-op-embed.js';
+import { buildCustomId, buildOpEmbed } from '../../../discord/utils/index.js';
 const EMOJIS: Array<string> = ['🍖', '🏴‍☠️', '⚓', '🐉', '🗺️', '🛶', '☠️', '🍈', '🦈', '🐟', '🏝️', '⚔️', '🔫', '🔪'];
 
 export const emojiCommand: Command = {
