+++ apps/bot/src/domains/_dev/commands/moi.ts
@@ -1,6 +1,5 @@
 import type { Command } from '../../../discord/types.js';
-import { buildOpEmbed } from '../../../discord/utils/build-op-embed.js';
-import { getTargetUser } from '../../../discord/utils/get-target-user.js';
+import { buildOpEmbed, getTargetUser } from '../../../discord/utils/index.js';
 
 // TODO: supprimer avant la prod
 export const moiCommand: Command = {
