+++ apps/bot/src/domains/_dev/interactions/randomcat-button.ts
@@ -1,7 +1,7 @@
 import type { ButtonInteraction } from 'discord.js';
 
 import type { ButtonHandler } from '../../../discord/types.js';
-import { buildOpEmbed } from '../../../discord/utils/build-op-embed.js';
+import { buildOpEmbed } from '../../../discord/utils/index.js';
 
 export const randomCatButtonHandler: ButtonHandler = {
   name: 'cat',
