+++ apps/bot/src/domains/resource/commands/inventaire.ts
@@ -1,5 +1,5 @@
 import type { Command } from '../../../discord/types.js';
-import { getTargetUser } from '../../../discord/utils/get-target-user.js';
+import { getTargetUser } from '../../../discord/utils/index.js';
 import { findOrCreatePlayer } from '../../player/service.js';
 import { buildInventoryView } from '../inventory-view.js';
 import { getInventory } from '../repository.js';
